requirements.txt
absl-py==2.3.1
asciitree==0.3.3
astunparse==1.6.3
attrs==25.3.0
crc32c==2.7.1
Deprecated==1.2.18
dm-tree==0.1.9
donfig==0.8.1.post1
fasteners==0.19
filelock==3.13.1
fsspec==2024.6.1
gast==0.6.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numcodecs==0.15.1
numpy==2.3.2
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-dali-cuda120==1.50.0
nvidia-nccl-cu12==2.26.2
nvidia-nvcomp-cu12==4.2.0.14
nvidia-nvimgcodec-cu12==0.5.0.13
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvjpeg-cu12==12.4.0.76
nvidia-nvjpeg2k-cu12==0.8.1.40
nvidia-nvtiff-cu12==0.5.0.67
nvidia-nvtx-cu12==12.8.55
packaging==24.2
pandas==2.3.1
pillow==11.0.0
pyarrow==21.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
setuptools==70.2.0
six==1.17.0
sympy==1.13.3
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
triton==3.3.1
typing_extensions==4.14.1
tzdata==2025.2
wheel==0.45.1
wrapt==1.17.2
zarr==3.1.1
